add : pay point 모듈 생성

+++ backend/auth/src/main/java/com/paymong/auth/entity/PayPoint.java
@@ -20,5 +20,5 @@ public class PayPoint {
     private Long memberId;
 
     @Column(name = "point")
-    private Long point;
+    private Integer point;
 }
